restaurant_backend/requirements.txt
apturl==0.5.2
asgiref==3.6.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==4.0.1
Django==4.1.7
django-admin==2.0.2
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-excel-response2==3.0.5
django-rest-auth==0.9.5
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
duplicity==0.8.21
excel-base==1.0.4
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isoweek==1.3.3
jeepney==0.7.1
kazam==1.5.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.6.2
protobuf==3.12.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyDrive==1.3.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.8
screen==1.0.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.3
systemd-python==234
TimeConvert==3.0.13
tzdata==2023.2
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.17.1
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
